requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser
crayons==0.4.0
enum34==1.1.10
execnet==1.9.0
Faker
idna==3.2
iniconfig==1.1.1
lxml
namedlist==1.8
numpy
opencv-python
packaging
Pillow
pluggy
py
pyparsing
pytest
pytest-forked
pytest-ordering
pytest-sugar
pytest-xdist
python-dateutil
requests
selenium
six
termcolor
text-unidecode
toml
urllib3
webdriver-manager
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
enum34==1.1.10
execnet==1.9.0
Faker==8.11.0
idna==3.2
iniconfig==1.1.1
lxml==4.6.3
namedlist==1.8
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-emoji==0.2.0
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-sugar==0.9.4
pytest-xdist==2.3.0
python-dateutil==2.8.2
requests==2.26.0
selenium==3.141.0
six==1.16.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.4.2
allure-pytest==2.9.43
allure-python-commons==2.9.43
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
enum34==1.1.10
execnet==1.9.0
Faker==8.11.0
idna==3.2
iniconfig==1.1.1
lxml==4.6.3
namedlist==1.8
numpy==1.21.1
opencv-python==4.5.3.56
packaging==21.0
Pillow==8.3.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-emoji==0.2.0
pytest-forked==1.3.0
pytest-ordering==0.6
pytest-rerunfailures==10.1
pytest-sugar==0.9.4
pytest-xdist==2.3.0
python-dateutil==2.8.2
PyYAML==5.4.1
redis==3.5.3
requests==2.26.0
selenium==3.141.0
six==1.16.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
webdriver-manager==3.4.2
